+++ src/KokkosComm/mpi/alltoall.hpp
@@ -24,7 +24,7 @@
 
 #include "impl/pack_traits.hpp"
 #include "impl/types.hpp"
-#include "impl/error-handling.hpp"
+#include "impl/error_handling.hpp"
 
 namespace KokkosComm::Impl {
