+++ src/KokkosComm/mpi/send.hpp
@@ -25,7 +25,7 @@
 
 #include "impl/pack_traits.hpp"
 #include "impl/types.hpp"
-#include "impl/error-handling.hpp"
+#include "impl/error_handling.hpp"
 
 namespace KokkosComm::mpi {
